Rock solid debug() and kprintf() - Removed bloated code from printing routines

+++ kernel/int/idt.h
@@ -4,33 +4,32 @@
 #include <stdint.h>
 #include <stddef.h>
 #include "../common.h"
-#include "../drivers/screen/monitor.h"
+#include "../drivers/vga/vga.h"
 
 typedef struct regs {
-    long long int r15;
-    long long int r14;
-    long long int r13;
-    long long int r12;
-    long long int r11;
-    long long int r10;
-    long long int r9;
-    long long int r8;
-    long long int rsi;
-    long long int rdi;
-    long long int rbp;
-    long long int rdx;
-    long long int rcx;
-    long long int rbx;
-    long long int rax;
-    long long int isr_number;
-    long long int error_code;
-    long long int rip;
-    long long int cs; 
-    long long int rflags; 
-    long long int rsp;
-    long long int ss;
+    int64_t r15;
+    int64_t r14;
+    int64_t r13;
+    int64_t r12;
+    int64_t r11;
+    int64_t r10;
+    int64_t r9;
+    int64_t r8;
+    int64_t rsi;
+    int64_t rdi;
+    int64_t rbp;
+    int64_t rdx;
+    int64_t rcx;
+    int64_t rbx;
+    int64_t rax;
+    int64_t isr_number;
+    int64_t error_code;
+    int64_t rip;
+    int64_t cs; 
+    int64_t rflags; 
+    int64_t rsp;
+    int64_t ss;
 } regs_t;
-// typedef struct regs regs_t;
 
 struct idt_desc {
     uint16_t offset_low;  // offset bits 0..15
